とりあえずDBアクセス動いた
