+++ testapp/__init__.py
@@ -5,6 +5,5 @@ app = Flask(__name__)
 app.config.from_object('testapp.config')
 
 db = SQLAlchemy(app)
-from .models import bukken
 
 import testapp.views
